requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning (PINNED TO TRAINING ENVIRONMENT)
scikit-learn==1.6.1
pandas==2.3.1
numpy==2.2.6
joblib==1.3.2
xgboost==2.1.4

# Database
pymongo==4.9.1
motor==3.6.0

# PDF Generation
reportlab==4.2.5

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
